+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.0] - 2023-05-22
+
+### Changed
+
+- new Vanguard_AU downloader, using the details data. Shorter download.
+
 ## [2.5.0] - 2023-05-22
 
 ### Changed
 
-- new Vanguard_AU downloader
+- new Vanguard_AU downloader using the historical prices data.
 
 ### Security